Make setup interactive and restructure syncing as a command group
Decouple init from Git-based syncing so jean-claude can be used
solely for profile management. Syncing is now opt-in during init
and organized under a `sync` command group.

- `jean-claude init` creates local config without requiring Git,
  prompts whether to set up syncing (--sync/--no-sync flags for
  non-interactive use)
- New `jean-claude sync` command group: setup, push, pull, status
- Top-level push/pull/status deprecated with redirect messages
- Fix git push to use -u for proper upstream tracking on first push

Closes #13
