+++ src/commands/index.ts
@@ -3,3 +3,4 @@ export { pullCommand } from './pull.js';
 export { pushCommand } from './push.js';
 export { statusCommand } from './status.js';
 export { profileCommand } from './profile.js';
+export { syncCommand } from './sync.js';
